\name{copyTermFreq}
\alias{copyTermFreq}
\title{Copy term frequencies table to clipboard}
\description{Copy a term frequencies table to the clipboard.}
\details{This command copies a HTML version of a previously computed term
         frequencies table to the clipboard, to allow pasting it into e.g.
         a word processor.
}
\seealso{\code{\link{docTermFreqDlg}}, \code{\link{varTermFreqDlg}}, \code{\link{HTML2clip}} }
