\name{freqTermsDlg}
\alias{freqTermsDlg}
\title{List most frequent terms of a corpus}
\description{List terms with the highest number of occurrences in the document-term matrix of a corpus.}
\details{This dialog allows printing the most frequent terms of the corpus, which correspond
         to the columns with the highest total in the document-term matrix.
}

\seealso{\code{\link{DocumentTermMatrix}}, \code{\link{restrictTermsDlg}}, \code{\link{listTerms}},
         \code{\link{termsAssocDlg}} }
