\name{inspectCorpus}
\alias{inspectCorpus}
\title{Inspect corpus}
\description{See contents of all documents in the corpus.}
\details{This function opens a window with the contents of all documents in the current
         corpus. Note that the texts are shown as they were on import, i.e. before
         the processing steps (removing case, punctuation, numbers and stopwords, or
         stemming), which make the texts hard to read. Though, if the corpus was split,
         created chunks are shown separately.
}

\seealso{ \code{\link{importCorpusDlg}} }
