\name{freqTermsDlg}
\alias{freqTermsDlg}
\title{List most frequent terms of a corpus}
\description{List terms with the highest number of occurrences in the document-term matrix of a corpus.}
\details{This dialog allows printing the most frequent terms of the corpus, which correspond
         to the columns with the highest total in the document-term matrix. If a variable is chosen,
         the returned terms correspond to those with the highest total among the documents within each level
         of the variable. Number of occurrences are reported together with the terms in the table.
}

\seealso{\code{\link{DocumentTermMatrix}}, \code{\link{restrictTermsDlg}}, \code{\link{listTerms}},
         \code{\link{termsAssocDlg}} }
